Convert integration test to benchmark and optimize predictions
